Updated grapple script with grapple hooks as reusable scenes
Made improvements to the grapple power to improve feel and properly grapple enemies. Also changed hooks to be a scene with the ability to create moving hooks which carry the player with them(some weird stuff happens if the speed is to fast. easily fixable probably if it becomes a relevant issue).

Also some goofy physics stuff as a result of the changes that can be used for maybe more advanced platforming. Still good feel